Fix lost-update bug
The code `timeout microsecondsTimeout readNextSetting` could time out just
after `readNextSetting` had read (and consumed) the next setting, causing it to
be dropped. In practice since the timeout just elapsed we were about to go off
anyway, but if there was a leap second in play then we might have waited for a
bit longer than needed.